+++ packages/gamut-styles/src/__tests__/GamutProvider.test.tsx
@@ -62,7 +62,7 @@ describe(GamutProvider, () => {
       ),
     });
 
-    screen.getByText(JSON.stringify(theme));
+    screen.getByText(JSON.stringify({ ...theme, useLogicalProperties: false }));
   });
   it('it can have another GamutProvider as a child with creating multiple caches or globals', () => {
     renderView({