+++ packages/gamut-styles/src/typings/theme.d.ts
@@ -2,5 +2,7 @@ import '@emotion/react';
 import { CoreTheme } from '../themes';
 
 declare module '@emotion/react' {
-  export interface Theme extends CoreTheme {}
+  export interface Theme extends CoreTheme {
+    useLogicalProperties?: boolean;
+  }
 }