+++ packages/gamut-icons/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.57.0](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-icons@9.56.4...@codecademy/gamut-icons@9.57.0) (2026-03-17)
+
+### Features
+
+- **icon:** Adds MiniCalendarIcon ([f93a00a](https://github.com/Codecademy/gamut/commit/f93a00af990f869cb0e30bb1b894c9307a05cd3a))
+
+### [9.56.4](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-icons@9.56.3...@codecademy/gamut-icons@9.56.4) (2026-03-17)
+
+**Note:** Version bump only for package @codecademy/gamut-icons
+
+### [9.56.3](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-icons@9.56.2...@codecademy/gamut-icons@9.56.3) (2026-03-12)
+
+**Note:** Version bump only for package @codecademy/gamut-icons
+
+### [9.56.2](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-icons@9.56.1...@codecademy/gamut-icons@9.56.2) (2026-03-04)
+
+**Note:** Version bump only for package @codecademy/gamut-icons
+
 ### [9.56.1](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-icons@9.56.0...@codecademy/gamut-icons@9.56.1) (2026-03-02)
 
 **Note:** Version bump only for package @codecademy/gamut-icons